requirements.txt
aiohttp
aiosignal
async-timeout
attrs
beautifulsoup4
bs4
charset-normalizer
colorama
emoji
frozenlist
idna
loguru
lxml
multidict
Pillow
pyaes
Pyrogram
PySocks
PyYAML
soupsieve
TgCrypto
win32-setctime
yarl
pymtts
